app/requirements.txt
elevenlabs
speechrecognition
pyaudio
openai
dotenv
langgraph
grandalf
uvicorn
fastapi
pydantic
pydub
langsmith
langchain_core
google-generativeai